- Got rid of the TKFile ambuigity (both the volutil and the auth2 tokenfile   names were stored in the variable TKFile). - reading keys by doing a normal `memset(key,0,RPC2_KEYSIZE);   read(fd,key,RPC2_KEYSIZE);' instead of the various fscanf/fgets without   checking arraybounds or pre-initializing the buffer etc. Now we should   actually be done with that silly 8 character token bug.

+++ coda-src/venus/fso_cfscalls0.cc
@@ -78,14 +78,14 @@ void fsobj::FetchProgressIndicator(long offset)
     else if (offset == 0)           { last = 100; curr = 0; }
     else {
 	last = (stat.GotThisData * 100) / stat.Length;
-	curr = (offset * 100) / stat.Length;
+	curr = ((unsigned long)offset * 100) / stat.Length;
     }
 
     if (last != curr) {
 	MarinerLog("progress::fetching (%s) %lux\n", comp, curr);
     }
 
-    stat.GotThisData = offset;
+    stat.GotThisData = (unsigned long)offset;
 }
 
 int fsobj::Fetch(vuid_t vuid) {
